// frontend/src/interfaces/auth/IProfile.ts
export interface IProfile {
  profile_picture: string | null;
  profile_banner: string | null;
  biography: string;
  birthday: string;
  website: string;
  facebook: string;
  twitter: string;
  instagram: string;
  github: string;
  linkedin: string;
  youtube: string;
  tiktok: string;
  snapchat: string;
}